Do not install debug symbols
Labs uses a handbuilt nginx-extras package, which doesn't have
a -dbg installed (this slipped through during testing because
of weird interactions between labsdebrepo and this). Remove
debug symbols for now.

Labs proxies are going to switch to trusty soon enough, at which
point we will not need the hand built packages at all. Will
add this package back when that happens

Change-Id: Ib9cc687cdc4181f14eaaf9a5a3ae315fbd7438fc